+++ src/common/js/constants.js
@@ -33,7 +33,7 @@ Object.keys( Editor.C.Mode ).every(( k ) => {
 // internal (non exported) constants
 
 // the cookie name for the switch state, defined here to be available anywhere
-COOKIE_RESPONSIBLE = 'pwix:Editor';
+COOKIE_RESPONSIBLE = 'pwix:editor';
 COOKIE_SWITCH_STATE = 'switch_state';
 COOKIE_SWITCH_STATE_IDENTIFIER = COOKIE_RESPONSIBLE + '/' + COOKIE_SWITCH_STATE + '/';